+++ Estructura.Software.Web.API/Estructura.Software.Web.API.csproj
@@ -10,4 +10,8 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.4.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\Estructura.Software.Web.Application\Estructura.Software.Web.Application.csproj" />
+  </ItemGroup>
+
 </Project>
