+++ Estructura.Software.Web.Application/Application/Services/Users/CreateUserService.cs
@@ -4,14 +4,12 @@ using Estructura.Software.Web.API.Domain.Interfaces.Services.Users;
 
 namespace Estructura.Software.Web.API.Application.Services.Users
 {
-    public class DeleteUserService : IDeleteUserServices
+    public class CreateUserService : ICreateUserService
     {
         private readonly IUserRepository _userRepository;
-        public DeleteUserService(IUserRepository userRepository) => _userRepository = userRepository;
 
-        public User delete(string name)
-        {
-            return _userRepository.Delete(name);
-        }
+        public CreateUserService(IUserRepository userRepository) => _userRepository = userRepository;
+
+        public User Create(User user)=> _userRepository.Create(user);
     }
 }
