+++ Estructura.Software.Web.Application/Application/Services/Users/DeleteUserService.cs
@@ -1,16 +1,17 @@
 ﻿using Estructura.Software.Web.API.Domain.Entities;
 using Estructura.Software.Web.API.Domain.Interfaces.Repositories;
 using Estructura.Software.Web.API.Domain.Interfaces.Services.Users;
-using Microsoft.AspNetCore.DataProtection.Repositories;
 
 namespace Estructura.Software.Web.API.Application.Services.Users
 {
-    public class CreateUserServices : ICreateUserServices
+    public class DeleteUserService : IDeleteUserService
     {
         private readonly IUserRepository _userRepository;
+        public DeleteUserService(IUserRepository userRepository) => _userRepository = userRepository;
 
-        public CreateUserServices(IUserRepository userRepository) => _userRepository = userRepository;
-
-        public User Create(User user)=> _userRepository.Create(user);
+        public User delete(string name)
+        {
+            return _userRepository.Delete(name);
+        }
     }
 }
