+++ Estructura.Software.Web.Application/Interfaces/IScopeWhoIAm.cs
