+++ Estructura.Software.Web.Application/Interfaces/ISingletonWhoIAm.cs
