+++ Estructura.Software.Web.Application/Interfaces/ITransientWhoIAm.cs
