+++ Estructura.Software.Web.Domain/Domain/Entities/User.cs
