+++ Estructura.Software.Web.Persistence/Infrastructure/Interfaces/IRepositories/IUserRepository.cs
