+++ Estructura.Software.Web.Persistence/Infrastructure/Interfaces/IServices/IUsers/IGetAllUserService.cs
